rename files and resturcture metric functions
